build: refractoring and cleaning
