+++ game_of_greed/scores.py
@@ -1,0 +1,38 @@
+scores = {
+    (1, 1): 100,
+    (1, 2): 200,
+    (1, 3): 1000,
+    (1, 4): 2000,
+    (1, 5): 3000,
+    (1, 6): 4000,
+    (2, 1): 0,
+    (2, 2): 0,
+    (2, 3): 200,
+    (2, 4): 400,
+    (2, 5): 600,
+    (2, 6): 800,
+    (3, 1): 0,
+    (3, 2): 0,
+    (3, 3): 300,
+    (3, 4): 600,
+    (3, 5): 900,
+    (3, 6): 1200,
+    (4, 1): 0,
+    (4, 2): 0,
+    (4, 3): 400,
+    (4, 4): 800,
+    (4, 5): 1200,
+    (4, 6): 1600,
+    (5, 1): 50,
+    (5, 2): 100,
+    (5, 3): 500,
+    (5, 4): 1000,
+    (5, 5): 1500,
+    (5, 6): 2000,
+    (6, 1): 0,
+    (6, 2): 0,
+    (6, 3): 600,
+    (6, 4): 1200,
+    (6, 5): 1800,
+    (6, 6): 2400,
+}
