+++ tests/version_2/test_sim_basic.py
@@ -1,7 +1,7 @@
 import pytest
 from tests.flo import diff
 #from game_of_greed.game_logic import Game
-from game_of_greed.new_approach import Game
+from game_of_greed.game import Game
 
 pytestmark = [pytest.mark.version_2]
